Computer Lab Reservation
This is a java computer lab reservation where you can display an array of computers that are either in use, not in use, or non-working. You are able to choose if you want to display the array of computers, check how many are available, reserve a computer, or report a computer as non-operational.